backend/requirements.txt
fastapi
uvicorn
python-multipart
azure-storage-blob
openai
pdfplumber
python-dotenv